adding in some unit testing

+++ src/main/java/ie/cit/cloud/tickets/model/performance/EventCreator.java
@@ -1,7 +1,5 @@
 package ie.cit.cloud.tickets.model.performance;
 
-import java.util.Date;
-
 /**
  * this object is a helper object for the REST interface to accommodate the creation of
  * a new Event object.  The Event constructor has the ability to create Events.  This
@@ -16,27 +14,20 @@ public class EventCreator
 
 	private String locationName;
 
-//	private Date date;
-
 	private String eventName;
 
-	private Long ticketCount;
-
-//	private Long ticketPrice;
+	private Integer ticketCount;
 
 	public EventCreator()
 	{
 	}
 
-//	public EventCreator(final String performerName, final String locationName, final Date date, final String eventName, final Long ticketCount, final Long ticketPrice)
-	public EventCreator(final String performerName, final String locationName, final String eventName, final Long ticketCount)
+	public EventCreator(final String performerName, final String locationName, final String eventName, final Integer ticketCount)
 	{
 		this.performerName = performerName;
 		this.locationName = locationName;
-//		this.date = date;
 		this.eventName = eventName;
 		this.ticketCount = ticketCount;
-//		this.ticketPrice = ticketPrice;
 	}
 
 	public String getPerformerName()
@@ -49,25 +40,15 @@ public class EventCreator
 		return locationName;
 	}
 
-//	public Date getDate()
-//	{
-//		return date;
-//	}
-
 	public String getEventName()
 	{
 		return eventName;
 	}
 
-	public Long getTicketCount()
+	public int getTicketCount()
 	{
 		return ticketCount;
 	}
-
-//	public Long getTicketPrice()
-//	{
-//		return ticketPrice;
-//	}
 
 	public void setPerformerName(String performerName)
 	{
@@ -79,23 +60,13 @@ public class EventCreator
 		this.locationName = locationName;
 	}
 
-//	public void setDate(Date date)
-//	{
-//		this.date = date;
-//	}
-
 	public void setEventName(String eventName)
 	{
 		this.eventName = eventName;
 	}
 
-	public void setTicketCount(Long ticketCount)
+	public void setTicketCount(int ticketCount)
 	{
 		this.ticketCount = ticketCount;
 	}
-
-//	public void setTicketPrice(Long ticketPrice)
-//	{
-//		this.ticketPrice = ticketPrice;
-//	}
 }
